screencast: fix pipewire default id_node
The default node_id is SPA_ID_INVALID.

We are casting pipewire uint32_t node_id to int for printing since they are
currently staying in the range of low integer numbers. This makes
spotting an uninitialized node_id (casted to -1) much easier. Should be
corrected if that becomes an issue in the future.

+++ src/screencast/pipewire_screencast.c
@@ -111,7 +111,7 @@ static void pwr_handle_stream_state_changed(void *data,
 
 	logprint(INFO, "pipewire: stream state changed to \"%s\"",
 		pw_stream_state_as_string(state));
-	logprint(INFO, "pipewire: node id is %d", cast->node_id);
+	logprint(INFO, "pipewire: node id is %d", (int)cast->node_id);
 
 	switch (state) {
 	case PW_STREAM_STATE_STREAMING:
